task2/requirements.txt
torch
torchvision
transformers
datasets
accelerate
numpy
pandas
scikit-learn
Pillow
ipykernel
tqdm